refactor: LComponent명 및 폴더명 변경
- LogPitcher -> LogHitter
- ContextB -> Context
- min27604/baseball/#15
