--- FE/react-baseball/src/components/provider/ContextB.jsx
@@ -1,13 +1,0 @@
-import { createContext } from 'react';
-
-const boardHistory = {
-  S: 0,
-  B: 0,
-  O: 0,
-  H: 0,
-  HitInfo: ' ',
-};
-
-const BoardHistoryContext = createContext();
-
-export { boardHistory, BoardHistoryContext };
